requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.2
pyyaml>=6.0
lightgbm>=4.0
matplotlib>=3.9.0

